requirements.txt
aiohttp==3.9.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
PyJWT==2.8.0
python-dotenv==1.0.0
python-http-client==3.3.7
requests==2.31.0
sendgrid==6.10.0
starkbank-ecdsa==2.2.0
twilio==8.10.2
urllib3==2.1.0
yarl==1.9.3
